% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-series.R
\name{build_series}
\alias{build_series}
\title{Builds Multi-year Series for the Station Variables.}
\usage{
build_series(station)
}
\arguments{
\item{station}{character string given the WIGOS compatible station
identifier.}
}
\description{
For each of the variables of a given station, rbinds the anual data frames in
the C3S-QC format to make a series of several years.
}
