% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv-chl-data.R
\name{conv_chl_data}
\alias{conv_chl_data}
\title{Converts the Clilean Surface Records 1950-1958 to C3S-QC Format.}
\usage{
conv_chl_data()
}
\description{
Converts the Clilean Surface Records 1950-1958 to C3S-QC Format.
}
