% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-dd-16wcr2deg.R
\name{convert_dd_16wcr2deg}
\alias{convert_dd_16wcr2deg}
\title{Converts Wind Direction from 16-wind Compass Rose Points to Degrees.}
\usage{
convert_dd_16wcr2deg(dd16wcr)
}
\arguments{
\item{dd16wcr}{Vector with the wind direction values in 16-wind compass rose
points.}
}
\description{
Converts Wind Direction from 16-wind Compass Rose Points to Degrees.
}
