% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-w-kn2ms-1.R
\name{convert_w_kn2ms}
\alias{convert_w_kn2ms}
\title{Converts Wind Speed from Knots to Meters per Second.}
\usage{
convert_w_kn2ms(wkn)
}
\arguments{
\item{wkn}{Vector with the wind speed values in knots.}
}
\description{
Converts Wind Speed from Knots to Meters per Second.
}
