% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form-chl2qc.R
\name{form_chl2qc}
\alias{form_chl2qc}
\title{Converts to WP3 C3S-QC Format the Digitisations of Chilean Surface Records
1950-1958 performed by FCiências.ID.}
\usage{
form_chl2qc(digt)
}
\arguments{
\item{digt}{A data frame with the following seventeen columns:
\strong{station | dayr | year | month | day | hour | tdw | cloud | wdir
| wsp | ppa | tta | rrr1 | rrr2 | tmin | tmax | rhum}.}
}
\description{
Splits the annual digitisations per station into several data frames by type
of meteorological variable. The output data frames and text files are in the
WP3 C3S-QC format: \strong{variable code | year (YYYY) | month (MM) | day
(DD) | time (HHMM) | observation}.
}
\details{
\strong{Input:}
\itemize{
\item The output data frame of \code{\link{read_chl_ffcul}} with seventeen
columns: WIGOS compatible station ID, day of the year, year, month, day,
hour, dew point temperature, cloud cover, wind direction, wind speed, air
pressure, air temperature, accumulated precipitation at first hour,
accumulated precipitation at second hour, daily minimum temperature, daily
maximum temperature, relative humidity.
}
\strong{Output:}
\itemize{
\item A .RData and a .txt file without header for each one of the ten
variables digitised (td, n, dd, w, p, ta, rr, Tn, Tx, rh) -
'VariableCode_StationName_Year'. If misses the output for some variable it's
because doesn't exist any observations of that variable in the record.
\item A .txt file with the wind speed and another with the wind direction in
the original units - 'VariableCode_StationName_Year_16wcr' and
'VariableCode_StationName_Year_kt'.
\item A .txt file for each one of the ten variables digitised which includes
the missing values in the anual series - 'VariableCode_StationName_Year_all'.
}
}
