% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-chl-ffcul.R
\name{read_chl_ffcul}
\alias{read_chl_ffcul}
\title{Reads the Digitisations of Clilean Surface Records 1950-1958 performed by
FCiências.ID.}
\usage{
read_chl_ffcul(obstxt)
}
\arguments{
\item{obstxt}{Character string giving the path of the metadata file.}
}
\value{
A data frame with seventeen columns: WIGOS compatible station ID, day
  of the year, year, month, day, hour, dew point temperature, cloud cover,
  wind direction, wind speed, air pressure, air temperature, accumulated
  precipitation at first hour, accumulated precipitation at second hour,
  daily minimum temperature, daily maximum temperature, relative humidity.
}
\description{
Reads one text file with the station metadata and twelve text files with the
digitisation tables for each month of a station-year. As all the data was
typed, performs the quality control of the station metadata and checks for
table format errors and typing errors.
}
\details{
\strong{Input:}
\itemize{
\item One text file with the metadata - 'StationName_Obs_YYYY.txt'. The
following constants are read from the metadata file:
\itemize{
\item WIGOS compatible station identifier
\item Observations year
\item Longitude
\item Latitude
\item Altitude
\item 'msl pressure' or 'surface pressure'
\item Real station name
}
\item Twelve text files with the digitisations, which correspond to a
digitisation table for each month of a station-year -
'StationName_Mth_YYYY.txt'. The digitisation tables have 13 columns with the
following variables:
\describe{
\item{Column 1}{Day of the observation}
\item{Column 2}{Observation time in the format HHMM}
\item{Column 3}{Dew point temperature in Celsius degrees (ºC)}
\item{Column 4}{Cloud cover in oktas: 0,1,2,3,4,5,6,7,8,9}
\item{Column 5}{Wind direction in 16-wind compass rose points, plus C - Calm}
\item{Column 6}{Wind speed in knots (kn, kt)}
\item{Column 7}{Atmospheric pressure - mean sea level or surface - in
hectopascals (hPa)}
\item{Column 8}{Air temperature in Celsius degrees (ºC)}
\item{Column 9}{Acumulated precipitation in milimeters (mm) most commonly
measured at 12:00 UTC}
\item{Column 10}{Acumulated precipitation in milimeters (mm) most commonly
measured at 23:00 UTC}
\item{Column 11}{Minimum daily temperature in Celsius degrees (ºC) most
commonly measured at 12:00 UTC}
\item{Column 12}{Maximum daily temperature in Celsius degrees (ºC) most
commonly measured at 23:00 UTC}
\item{Column 13}{Relative humidity in percent}
}
}
\strong{Output:}
\itemize{
\item A .RData and a .txt file with the metadata bellonging to a
station-year: 'metadata_StationName_Year'.
\item A .RData and a .txt file with the anual digitisation of station-year:
'digitisation_StationName_Year'.
\item Eventually, one or two .txt files with typing errors to be corrected
before proceeding with the quality control:
'Hour-errors_StationName_YYYY.txt' and 'Typing-errors_StationName_YYYY.txt'.
}
}
